[BBC-12729] Caption component is now internal to Review
