+++ src/index.tsx
@@ -5,7 +5,6 @@ export * from './breadcrumb'
 export * from './bullet'
 export * from './button'
 export * from './buttonGroup'
-export * from './caption'
 export * from './confirmationModal'
 export * from './datePicker'
 export * from './disclaimer'
